fix: FTA score sim-cff-event-adapter
